moved config back to API from constants to avoid conflicts
